+++ CHANGELOG.md
@@ -10,6 +10,8 @@ All notable changes to this project will be documented in this file.
 - Body: compute the adjoint matrix in ``body.adjoint_matrix``
 - Contact: ``contact.normal`` alias
 - Contact: ``wrench``, ``force`` and ``moment`` attributes
+- Contact: ``set_wrench()`` function that switches a contact to managed mode
+- Contact: ``wrench_at()`` function to get wrench at a given point in the world frame
 - Example: [stabilization with height variations](https://hal.archives-ouvertes.fr/hal-02289919v1/document)
 - InvertedPendulum: draw center of pressure by ``draw_point``
 - Manipulator: ``wrench``, ``force`` and ``moment`` attributes